chore: renames testbed config utility to prevent test errors
