+++ src/app/adf-user-management/df-register/df-register.component.spec.ts
@@ -1,6 +1,6 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { DfRegisterComponent } from './df-register.component';
-import { createTestBedConfig } from 'src/app/shared/utilities/test';
+import { createTestBedConfig } from 'src/app/shared/utilities/testbed-config';
 import { DfAuthService } from '../services/df-auth.service';
 import { DfSystemConfigDataService } from 'src/app/shared/services/df-system-config-data.service';
